refactor staking app
